refactor(imports): update import paths for User entity across multiple files
