+++ server/src/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { User } from "src/user/entities/user.entities";
+import { User } from "src/entities/user.entities";
 import { Repository } from "typeorm";
 import * as bcrypt from "bcrypt"
 import { JwtService } from "@nestjs/jwt";
